Separate view functional into different classes
